ISEN-198 Fix for silent CDXGen failure (#159)
* Fix for silent CDXGen failure

On certain projects, CDXGen fails silently, but still returns
exit code 0. As we assume that the program ran successfuly,
we attempt to open the generated SBOM file and fail.

The fix/refactor makes it so that failing to read a file is
considered a failure on a single command level, not the
entire attempt to generate. We will retry the generation
with no licences, and that should not fail directly.

* Adjust logging levels on errors

* Remove double import for cyclonecdx-go package

* Change FIXME to TODO